fix: InputDTO 삭제 및 InputView에 입력 관련 기능을 응집
